fix: logging pointless error when offline
