+++ src/extensions/replay/web-performance.ts
@@ -1,9 +1,9 @@
-import { logger } from '../../utils'
 import { PostHog } from '../../posthog-core'
 import { DecideResponse, NetworkRequest } from '../../types'
 import { isLocalhost } from '../../utils/request-utils'
 
 import { _isUndefined } from '../../utils/type-utils'
+import { logger } from '../../utils/logger'
 
 const PERFORMANCE_EVENTS_MAPPING: { [key: string]: number } = {
     // BASE_PERFORMANCE_EVENT_COLUMNS
