+++ src/posthog-persistence.ts
@@ -1,6 +1,6 @@
 /* eslint camelcase: "off" */
 
-import { _each, _extend, _include, _strip_empty_properties, logger } from './utils'
+import { _each, _extend, _include, _strip_empty_properties } from './utils'
 import { cookieStore, localStore, localPlusCookieStore, memoryStore, sessionStore } from './storage'
 import { PersistentStore, PostHogConfig, Properties } from './types'
 import {
@@ -13,6 +13,7 @@ import {
 
 import { _isObject, _isUndefined } from './utils/type-utils'
 import { _info } from './utils/event-utils'
+import { logger } from './utils/logger'
 
 const CASE_INSENSITIVE_PERSISTENCE_TYPES: readonly Lowercase<PostHogConfig['persistence']>[] = [
     'cookie',
