+++ src/retry-queue.ts
@@ -1,10 +1,10 @@
 import { RequestQueueScaffold } from './base-request-queue'
 import { encodePostData, xhr } from './send-request'
 import { QueuedRequestData, RetryQueueElement } from './types'
-import { logger } from './utils'
 import { RateLimiter } from './rate-limiter'
 
 import { _isUndefined } from './utils/type-utils'
+import { logger } from './utils/logger'
 
 const thirtyMinutes = 30 * 60 * 1000
