+++ src/send-request.ts
@@ -1,9 +1,10 @@
-import { _each, logger } from './utils'
+import { _each } from './utils'
 import Config from './config'
 import { PostData, XHROptions, XHRParams } from './types'
 import { _HTTPBuildQuery } from './utils/request-utils'
 
 import { _isArray, _isFunction, _isNumber, _isUint8Array, _isUndefined } from './utils/type-utils'
+import { logger } from './utils/logger'
 
 export const addParamsToURL = (
     url: string,
