+++ src/sessionid.ts
@@ -3,9 +3,10 @@ import { SESSION_ID } from './constants'
 import { sessionStore } from './storage'
 import { PostHogConfig, SessionIdChangedCallback } from './types'
 import { uuidv7 } from './uuidv7'
-import { logger, window } from './utils'
+import { window } from './utils'
 
 import { _isArray, _isNumber, _isUndefined } from './utils/type-utils'
+import { logger } from './utils/logger'
 
 const MAX_SESSION_IDLE_TIMEOUT = 30 * 60 // 30 minutes
 const MIN_SESSION_IDLE_TIMEOUT = 60 // 1 minute
