+++ src/storage.ts
@@ -1,8 +1,9 @@
-import { _extend, logger } from './utils'
+import { _extend } from './utils'
 import { PersistentStore, Properties } from './types'
 import { DISTINCT_ID, SESSION_ID, SESSION_RECORDING_IS_SAMPLED } from './constants'
 
 import { _isNull, _isUndefined } from './utils/type-utils'
+import { logger } from './utils/logger'
 
 const DOMAIN_MATCH_REGEX = /[a-z0-9][a-z0-9-]+\.[a-z]{2,}$/i
