+++ src/utils/index.ts
@@ -1,4 +1,3 @@
-import Config from '../config'
 import { Breaker, EventHandler, Properties } from '../types'
 import {
     _isArray,
@@ -10,6 +9,7 @@ import {
     _isUndefined,
     hasOwnProperty,
 } from './type-utils'
+import { logger } from './logger'
 
 /*
  * Saved references to long variable names, so that closure compiler can
@@ -23,44 +23,6 @@ const navigator = win.navigator || { userAgent: '' }
 const document = win.document || {}
 const userAgent = navigator.userAgent
 const breaker: Breaker = {}
-
-const LOGGER_PREFIX = '[PostHog.js]'
-
-export const logger = {
-    _log: (level: 'log' | 'warn' | 'error', ...args: any[]) => {
-        if ((Config.DEBUG || (window as any).POSTHOG_DEBUG) && !_isUndefined(window.console) && window.console) {
-            const consoleLog =
-                '__rrweb_original__' in window.console[level]
-                    ? (window.console[level] as any)['__rrweb_original__']
-                    : window.console[level]
-
-            // eslint-disable-next-line no-console
-            consoleLog(LOGGER_PREFIX, ...args)
-        }
-    },
-
-    info: (...args: any[]) => {
-        logger._log('log', ...args)
-    },
-
-    warn: (...args: any[]) => {
-        logger._log('warn', ...args)
-    },
-
-    error: (...args: any[]) => {
-        logger._log('error', ...args)
-    },
-
-    critical: (...args: any[]) => {
-        // Critical errors are always logged to the console
-        // eslint-disable-next-line no-console
-        console.error(LOGGER_PREFIX, ...args)
-    },
-
-    uninitializedWarning: (methodName: string) => {
-        logger.error(`You must initialize PostHog before calling ${methodName}`)
-    },
-}
 
 // UNDERSCORE
 // Embed part of the Underscore Library
